--- packages/package.json
@@ -1,46 +1,0 @@
-{
-    "name": "@alqmc/build",
-    "version": "0.0.1",
-    "description": "build 工具",
-    "author": "alqmc",
-    "license": "MIT",
-    "main": "./lib/index.js",
-    "module": "./es/index.mjs",
-    "publishConfig": {
-        "access": "public"
-    },
-    "exports": {
-        ".": {
-            "require": "./lib/index.js",
-            "import": "./es/index.mjs"
-        },
-        "./es": "./es/index.mjs",
-        "./lib": "./lib/index.js",
-        "./es/*.mjs": "./es/*.mjs",
-        "./es/*": "./es/*.mjs",
-        "./lib/*.js": "./lib/*.js",
-        "./lib/*": "./lib/*.js",
-        "./*": "./*"
-    },
-    "dependencies": {
-        "@rollup/plugin-commonjs": "^22.0.0",
-        "@rollup/plugin-json": "^4.1.0",
-        "@rollup/plugin-node-resolve": "^13.2.1",
-        "@ts-morph/common": "^0.13.0",
-        "@vitejs/plugin-vue": "^2.3.1",
-        "@vitejs/plugin-vue-jsx": "^1.3.10",
-        "archiver": "^5.3.1",
-        "chalk": "4.1.2",
-        "esbuild": "^0.14.38",
-        "fast-glob": "^3.2.11",
-        "fs-extra": "^10.1.0",
-        "gulp": "^4.0.2",
-        "prettier": "^2.6.2",
-        "rimraf": "^3.0.2",
-        "rollup": "^2.70.2",
-        "rollup-plugin-esbuild": "^4.9.1",
-        "ts-morph": "^14.0.0",
-        "typescript": "^4.6.3",
-        "vite": "^2.9.5"
-    }
-}
