--- packages/tasks/index.ts
@@ -1,4 +1,0 @@
-import { buildTypescriptLib } from './typescriptLib';
-import { buildVueLib } from './vueLib';
-
-export { buildTypescriptLib, buildVueLib };
