+++ packages/typescript/README.md
@@ -1,0 +1,3 @@
+## @alqmc/build-typescript
+
+typescript lib打包库
