+++ packages/utils/README.md
@@ -1,0 +1,3 @@
+## @alqmc/build-utils
+
+基础工具库
