+++ packages/utils/index.ts
@@ -3,4 +3,13 @@ import { log } from './src/log';
 import { updateVersion } from './src/version';
 import { createZip } from './src/buildZip';
 import { run, withTask } from './src/process';
-export { formatCode, log, updateVersion, createZip, run, withTask };
+import { getExternal } from './src/pkg';
+export {
+  formatCode,
+  log,
+  updateVersion,
+  createZip,
+  run,
+  withTask,
+  getExternal,
+};
