+++ packages/utils/package.json
@@ -1,28 +1,44 @@
 {
     "name": "@alqmc/build-utils",
-    "version": "0.0.1",
-    "description": "build typescript lib",
+    "version": "0.0.1-dev",
+    "description": "build utils lib",
     "author": "alqmc",
     "license": "MIT",
+    "main": "./lib/index.js",
+    "module": "./es/index.mjs",
+    "types": "types",
     "publishConfig": {
         "access": "public"
+    },
+    "exports": {
+        ".": {
+            "require": "./lib/index.js",
+            "import": "./es/index.mjs"
+        },
+        "./es": "./es/index.mjs",
+        "./lib": "./lib/index.js",
+        "./es/*.mjs": "./es/*.mjs",
+        "./es/*": "./es/*.mjs",
+        "./lib/*.js": "./lib/*.js",
+        "./lib/*": "./lib/*.js",
+        "./*": "./*"
     },
     "peerDependencies": {
         "typescript": "^4.6.3"
     },
     "devDependencies": {
+        "@pnpm/types": "^8.0.0",
+        "@types/archiver": "^5.3.1",
+        "@types/fs-extra": "^9.0.13",
         "@types/node": "^17.0.25",
         "@types/prettier": "^2.6.0",
-        "@types/archiver": "^5.3.1",
         "@types/gulp": "^4.0.9",
-        "@types/fs-extra": "^9.0.13",
-        "@pnpm/types": "^8.0.0",
         "gulp": "^4.0.2"
     },
     "dependencies": {
         "archiver": "^5.3.1",
-        "prettier": "^2.6.2",
         "chalk": "4.1.2",
-        "fs-extra": "^10.1.0"
+        "fs-extra": "^10.1.0",
+        "prettier": "^2.6.2"
     }
 }
