+++ packages/vue/package.json
@@ -4,8 +4,24 @@
     "description": "build vue lib",
     "author": "alqmc",
     "license": "MIT",
+    "types": "types",
+    "main": "./lib/index.js",
+    "module": "./es/index.mjs",
     "publishConfig": {
         "access": "public"
+    },
+    "exports": {
+        ".": {
+            "require": "./lib/index.js",
+            "import": "./es/index.mjs"
+        },
+        "./es": "./es/index.mjs",
+        "./lib": "./lib/index.js",
+        "./es/*.mjs": "./es/*.mjs",
+        "./es/*": "./es/*.mjs",
+        "./lib/*.js": "./lib/*.js",
+        "./lib/*": "./lib/*.js",
+        "./*": "./*"
     },
     "dependencies": {
         "@rollup/plugin-commonjs": "^22.0.0",
